# Stop the bleed on chunk mapper panic