--- a/tsdb/db_test.go
+++ b/tsdb/db_test.go
@@ -3127,6 +3127,8 @@ func TestNoPanicOnTSDBOpenError(t *testing.T) {
 }
 
 func TestQuerier_ShouldNotPanicIfHeadChunkIsTruncatedWhileReadingQueriedChunks(t *testing.T) {
+	t.Skip("TODO: investigate why the process crashes with no output when running in CI")
+
 	const (
 		numSeries                = 1000
 		numStressIterations      = 10000