This PR is a trivial fix to #8318 and #8217, which has been reproduced in a unit test in #8681 (unit test is also added in this PR).

The fix done in this PR is simply copying the chunk `[]byte` slice when reading it. It clearly has a negative impact on the memory utilization and allocations at query time but could allow us to gain some time before a more elaborate fix is ready (eg. see #8722 as an attempt).

_Could we run prombench on this PR to see how bad it is in terms of performances?_