@pracucci 2021-04-15
@codesome @roidelapluie I don't see a significative impact on query latency or memory. I'm not familiar with prombench: are mmap-ed head chunks stressed as well?