@roidelapluie 2021-04-18
> @codesome @roidelapluie I don't see a significative impact on query latency or memory. I'm not familiar with prombench: are mmap-ed head chunks stressed as well?

Well, we run queries against the head.