@pracucci 2021-04-21
I've tested it into a dev Cortex cluster, trying to stress test the mmap-ed head chunks (basically running 50K+ cardinality queries over -1h,-2h). I've also added a metric to count the extra bytes allocated by this copy (see [code here](https://github.com/cortexproject/cortex/compare/pracucci:test-prometheus-pr-8723)).

As a data point, I can see that, under this test scenario, the bytes allocated increase by about 5% (compare to the total bytes allocated on the Go heap).

![Screenshot 2021-04-21 at 18 22 42](https://user-images.githubusercontent.com/1701904/115588419-058d2c00-a2cf-11eb-83a3-e19afffdaf46.png)

The next chart shows:
- Left: MB/s allocated by the chunk copy
- Right: MB/s allocated by the whole Go heap

![Screenshot 2021-04-21 at 18 22 34](https://user-images.githubusercontent.com/1701904/115588411-032ad200-a2cf-11eb-9b0c-3d941e0234ee.png)