@codesome 2021-04-22
5% bytes allocation sounds better than panic. This should be a good temporary fix.