@pracucci 2021-04-22
However the test `TestQuerier_ShouldNotPanicIfHeadChunkIsTruncatedWhileReadingQueriedChunks` looks panicking in CI (can't reproduce the panic anymore in my dev env). I can't see the full `go test` output from the CI: could you help me getting it, please? I would like to understand why/where it's panicking.