@pracucci 2021-04-22
In Cortex we also have a couple of benchmarks on querying local ingesters TSDB via `tsdb.Querier()` and `tsdb.ChunkQuerier()`. I've modified `MaxHeadChunkFileSize`  from `128MB` to `128KB` in order to stress test the mmap-ed head chunks when querying and I get the following diff when chunks copy is enabled:

```
name                               old time/op    new time/op    delta
Ingester_v2QueryStream_Samples-12    5.92ms ± 0%    5.99ms ± 1%   +1.10%  (p=0.008 n=5+5)
Ingester_v2QueryStream_Chunks-12     1.27ms ±12%    1.19ms ± 0%     ~     (p=0.151 n=5+5)

name                               old alloc/op   new alloc/op   delta
Ingester_v2QueryStream_Samples-12    3.60MB ± 0%    3.83MB ± 0%   +6.32%  (p=0.008 n=5+5)
Ingester_v2QueryStream_Chunks-12      385kB ± 0%     499kB ± 0%  +29.59%  (p=0.008 n=5+5)

name                               old allocs/op  new allocs/op  delta
Ingester_v2QueryStream_Samples-12     8.34k ± 0%     9.68k ± 0%  +16.07%  (p=0.008 n=5+5)
Ingester_v2QueryStream_Chunks-12      5.20k ± 0%     5.87k ± 0%  +12.89%  (p=0.008 n=5+5)
```

This pretty match matches my expectations. Previously I've mentioned a +5% on memory allocations in Cortex ingesters when copy is enabled and that consider the facts that ingesters does way more than querying chunks (like in Prometheus the memory allocated by chunks copy is a fraction of the total memory allocated both by scraping and querying).

In the Cortex scenario, even if the copy, fetching chunks via `ChunkQuerier()` looks way more convenient than fetching samples via `Querier()`:
- Without copy: fetching chunks allocates -89% memory compared to fetching samples
- With copy: fetching chunks allocated -86% memory compared to fetching samples