@gouthamve 2021-04-22
Manually SSH-ed into the run and ran:

```
circleci@2926b34fcc29:~/project$ /usr/local/go/bin/go test ./tsdb -run=TestQuerier_ShouldNotPanicIfHeadChunkIsTruncatedWhileReadingQueriedChunks

signal: killed
FAIL    github.com/prometheus/prometheus/tsdb   5.941s
FAIL
circleci@2926b34fcc29:~/project$ 
circleci@2926b34fcc29:~/project$ /usr/local/go/bin/go test ./tsdb -run=TestQuerier_ShouldNotPanicIfHeadChunkIsTruncatedWhileReadingQueriedChunks
signal: killed
FAIL    github.com/prometheus/prometheus/tsdb   4.420s
FAIL
circleci@2926b34fcc29:~/project$ /usr/local/go/bin/go test ./tsdb -run=TestQuerier_ShouldNotPanicIfHeadChunkIsTruncatedWhileReadingQueriedChunks
signal: killed
FAIL    github.com/prometheus/prometheus/tsdb   5.096s
FAIL
circleci@2926b34fcc29:~/project$ 
```

Does it take too much memory?