@pracucci 2021-04-22
> Does it take too much memory?

I also suspect the memory pressure is too much. It should allocate about 1.5GB. GC runs on 100% from previous allocated memory, so the test process may test about 3GB. Maybe CI doesn't have much memory?

_I've just pushed a change to reduce the memory pressure. In my local machine it still panics without the fix._