@roidelapluie 2021-04-22
run dmesg