@codesome 2021-04-28
I have another (brilliant?) idea: Why not we just delay truncating the m-mapped chunks by some X minutes? 

Basically [here](https://github.com/prometheus/prometheus/blob/e36e5fa83399dd899fafa5dd09d09fa1d20a4f22/tsdb/head.go#L891), we can do `mint-delta`. The `delta` should be enough that within that time all queries which would be in flight would have finished.

This might mean that we will keep around the m-mapped files till next Head compaction happens, since truncate is only called at that time, hence more disk space. But there is a solution for that: Head block calls truncate on the m-mapped files at fixed intervals (say 15m) for the last truncation time attempted. This is not an expensive operation hence viable.

I am thinking of `delta` to be `15m`.