@roidelapluie 2021-04-28
We do not really need to guess the delta, we have a timeout for queries. I know we do not interrupt the queries after the timeout if they are in the TSDB phase but it would seem like a good first approach.