@codesome 2021-04-28
Since we cannot assume the time unit in TSDB, we can best to `delta = head.chunkRange/X`. So with assumption of chunk range as 2h, `X` could be 8. (If someone intends to reduce the size of blocks, then X=8 should fair well upto some extent)