@roidelapluie 2021-04-28
This is not a time unit we pass  and store inside the TSDB, this is the time we allow for a query to run. There is no "unit" involved, it is just the time that passes.