@pracucci 2021-04-29
> Basically here, we can do mint-delta. The delta should be enough that within that time all queries which would be in flight would have finished.

By "here" you linked to this:
```
h.chunkDiskMapper.Truncate(mint)
```

If you call `h.chunkDiskMapper.Truncate(mint - delta)` (whatever the delta is computed) it will truncate chunks whose highest sample timestamp is up to `mint - delta` but, to my understanding, it's perfectly legit there's a on-going query running on such data.

If you want to delay the truncation by X time, then what you should do is that `Truncate()` just marks head chunk files as "do not use" (which is the approach I've used in the experimental #8722) and then you will actually close the mmap-ed file and delete it after X time. This removes the complexity of having to track readers (see #8722) at the cost of having no strong guarantee that there's no active reader.