@codesome 2021-05-03
My code of delta was definitely wrong! Thanks for pointing out

My idea was something like 
```
defer func(){
    select {
    case <-time.After(15 * time.Minute):
        h.chunkDiskMapper.Truncate(mint)
    case <-someStopChannel:
    }
}()
```

And that should be enough for real world scenarios.