@pracucci 2021-05-05
> My idea was something like
> 
> ```
> defer func(){
>     select {
>     case <-time.After(15 * time.Minute):
>         h.chunkDiskMapper.Truncate(mint)
>     case <-someStopChannel:
>     }
> }()
> ```

I'm a bit dubious about this cause we loose control over it (eg. in tests), but I understand this hack could protect us.