@roidelapluie 2021-05-05
The way do deal with that in tests would be to have the timeout as a TSDB parameter, correct?