@codesome 2021-05-05
Yes this could be a configuration. Depending on how long queries are allowed by downstream users, this will help tweak the timeout here.